add map location on addstory

+++ lib/features/add_story_screen/data/models/upload_story.dart
@@ -1,11 +1,10 @@
-import 'dart:ffi';
 import 'package:image_picker/image_picker.dart';
 
 class UploadStory {
   final XFile file;
   final String description;
-  final Float? lat;
-  final Float? lon;
+  final double? lat;
+  final double? lon;
 
   UploadStory({
     required this.file,
